Move lims sample to models
